Commit From Erd

+++ beinplannercloud/beinplannercloud.service/src/main/java/tr/com/beinplanner/definition/repository/DefFirmRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import tr.com.beinplanner.definition.dao.DefFirm;
 
 @Repository
-public interface DefFirmRepository  extends CrudRepository<DefFirm, Long>{
+public interface DefFirmRepository  extends CrudRepository<DefFirm, Integer>{
 
 }
